Clean up autonomous code and have more concise namings
